requirements.txt
absl-py==0.7.0
astor==0.7.1
bz2file==0.98
certifi==2020.4.5.1
chardet==3.0.4
ci-info==0.1.1
click==7.1.1
cycler==0.10.0
decorator==4.4.2
-e git+https://github.com/marcoancona/DeepExplain.git@16fb0f298bc676318fe535c334cb429bbc3eefa0#egg=deepexplain
etelemetry==0.2.0
filelock==3.0.12
future==0.18.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.9
imageio==2.9.0
isodate==0.6.0
joblib==0.14.1
Keras==2.2.5
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.0
keras-vis==0.4.1
kiwisolver==1.2.0
lxml==4.5.0
Markdown==3.0.1
matplotlib==3.2.1
networkx==2.4
neurdflib==5.0.1
nibabel==2.3.3
nilearn==0.6.2
nipype==1.4.2
numexpr==2.7.1
numpy==1.16.4
packaging==20.3
pandas==1.0.3
Pillow==7.2.0
protobuf==3.6.1
prov==1.5.3
pydot==1.4.1
pydotplus==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
rdflib==4.2.2
requests==2.23.0
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.1.0
shap==0.35.0
SimpleITK==1.2.4
simplejson==3.17.0
six==1.14.0
sklearn==0.0
tables==3.6.1
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
tifffile==2020.8.13
tqdm==4.48.2
traits==6.0.0
urllib3==1.25.8
Werkzeug==1.0.1
xgboost==1.0.2
